Added extra API Functions; Minor design changes to database

+++ DAL/Entities/I_Permission.cs
@@ -8,9 +8,9 @@ namespace DAL.Entities
 {
     public class I_Permission
     {
-        public int I_Permission_ID { get; set; }
-        public int Image_ID { get; set; }
-        public int User_ID { get; set; }
+        public Int64 I_Permission_ID { get; set; }
+        public Int64 Image_ID { get; set; }
+        public Int64 User_ID { get; set; }
         public String? I_Permission_Type { get; set; }
 
         public Image? Image { get; set; } //BN Comment - why add this??
