+++ Gallery2/Models/A_Permission/A_Permission_Pass_Object.cs
@@ -2,8 +2,8 @@
 {
     public class A_Permission_Pass_Object
     {
-        public int album_id { get; set; }
-        public int user_id { get; set; }
+        public Int64 album_id { get; set; }
+        public Int64 user_id { get; set; }
         public String a_permission_type { get; set; }
     }
 }
