+++ Gallery2/Models/Album/AlbumUpdate_Pass_Object.cs
@@ -2,6 +2,6 @@
 {
     public class AlbumUpdate_Pass_Object: Album_Pass_Object
     {
-        public int album_id { get; set; }
+        public Int64 album_id { get; set; }
     }
 }
