+++ Gallery2/Models/I_Permission/I_Permission_Pass_Object.cs
@@ -2,8 +2,8 @@
 {
     public class I_Permission_Pass_Object
     {
-        public int image_id { get; set; }
-        public int user_id { get; set; }
+        public Int64 image_id { get; set; }
+        public Int64 user_id { get; set; }
         public String i_permission_type { get; set; }
     }
 }
