+++ LOGIC/Services/Models/A_Permission/A_Permission_ResultSet.cs
@@ -9,9 +9,9 @@ namespace LOGIC.Services.Models.A_Permission
 {
     public class A_Permission_ResultSet
     {
-        public int a_permission_id { get; set; }
-        public int album_id { get; set; }
-        public int user_id { get; set; }
+        public Int64 a_permission_id { get; set; }
+        public Int64 album_id { get; set; }
+        public Int64 user_id { get; set; }
         public String a_permission_type { get; set; }
 
        // public Album Album { get; set; } //BN comment - why?
